# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/tipopagamento": {
            "get": {
                "tags": [
                    "TipoPagamento"
                ],
                "summary": "Obtém todos os tipos de pagamento",
                "description": "Retorna uma lista com todos os tipos de pagamento cadastradas",
                "operationId": "getTipoPagamento",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de pagamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nome": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TipoPagamento"
                ],
                "summary": "Cria um novo tipo de pagamento",
                "description": "Cria um novo tipo de pagamento com o nome fornecido",
                "operationId": "createTipoPagamento",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de Pagamento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nome": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados fornecidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}